Move normalise from validation to transformers
